requirements.txt
kurigram
motor
dnspython
python-dotenv
beautifulsoup4
cloudscraper
requests
loguru
aiohttp
aiofiles
pypdf
Pillow
apscheduler

